TRY PINTURILLO PAINT/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.1.0
opencv-python==4.8.1.78
mediapipe==0.10.8
flask-socketio==5.3.5
python-socketio==5.10.0
